GPTQ updates
Summary:

1) reorganized GPTQ
 a) got rid of old GPTQ and renamed GPTQ_MT to GPTQ
 b) moved new GPTQ to prototype
 c) moved quantized linear modules in GPTQ.py to linear_quant_modules.py
2) removed dependence on lm_eval for input_recorder
 a) created new input recorder that doesn't depend on lm_eval
 b) made lm_eval input recorder depend on new generic input_recorder
 c) made TransformerEvalWrapper the base class and made
 d) updated apis generally to work with new input recorder
 LMEvalInputRecorder inherit from it instead of vice-versa
3) reorganized GPTQ tests
 a) moved tests from test_quant_api.py to test_gptq.py
 b) added new test that can be run in CI that doesn't depend on
 lm_eval/llama weights
 c) got rid of test_gptq_mt.py
4) added new documentation for lm_eval
5) GPTQ improvements
 a) reimplemented faster quant
 b) tested compilation of hessian calculation and parts of faster quant,
 generally they were slower.
 c) moved helper functions out of the class. They're largely generic and
 this is less cluttered.
 d) some improvements to the duplication checking and copying to be
 faster when possible
 e) fixed some bugs due to this not being in CI and things changing for
 int4wo tensor subclass.

Test Plan:

1) `python test_gptq.py`

note: the skipped test test_gptq_quantizer_int4_weight_only also ran.

2) I verified that all activation match between old GPTQ and current
   GPTQ

3)

```shell

export CHECKPOINT_PATH=../../../checkpoints # path to checkpoints folder

export MODEL_REPO=meta-llama/Llama-2-7b-chat-hf
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
--quantization int4wo-64
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
--quantization int4wo-gptq-64 --calibration_limit 10

export MODEL_REPO=meta-llama/Meta-Llama-3-8B
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
--quantization int4wo-64
python eval.py --checkpoint_path $CHECKPOINT_PATH/$MODEL_REPO/model.pth
--quantization int4wo-gptq-64 --calibration_limit 10

```
see README.md for results but they show GPTQ is working

Reviewers:

Subscribers:

Tasks:

Tags:
