+++ test/quantization/test_qat.py
@@ -17,12 +17,15 @@ from parameterized import parameterized
 from torch.ao.quantization.fx._decomposed import quantized_decomposed_lib  # noqa: F401
 
 from torchao import quantize_
-from torchao.quantization.GPTQ import _replace_linear_8da4w, _replace_linear_int4
 from torchao.quantization.granularity import (
     PerAxis,
     PerGroup,
     PerRow,
     PerToken,
+)
+from torchao.quantization.linear_quant_modules import (
+    _replace_linear_8da4w,
+    _replace_linear_int4,
 )
 from torchao.quantization.qat.api import (
     ComposableQATQuantizer,
