+++ torchao/_models/llama/generate.py
@@ -629,13 +629,13 @@ def main(
                     float8_dynamic_activation_float8_weight(granularity=granularity),
                 )
         elif "autoquant_v2" in quantization:
-            from torchao._models._eval import InputRecorder
+            from torchao._models._eval import LMEvalInputRecorder
             from torchao._models.llama.model import prepare_inputs_for_model
             from torchao.prototype.quantization.autoquant_v2 import autoquant_v2
 
             calibration_seq_length = 256
             inputs = (
-                InputRecorder(
+                LMEvalInputRecorder(
                     tokenizer,
                     calibration_seq_length,
                     prepare_inputs_for_model,
@@ -647,7 +647,7 @@ def main(
                     ["wikitext"],
                     1,
                 )
-                .get_inputs()[0]
+                .get_recorded_inputs()[0]
                 .values[0]
             )
             inputs = prepare_inputs_for_model(inputs)
@@ -719,12 +719,12 @@ def main(
             # do autoquantization
             model.finalize_autoquant()
         elif "autoquant" in quantization:
-            from torchao._models._eval import InputRecorder
+            from torchao._models._eval import LMEvalInputRecorder
             from torchao._models.llama.model import prepare_inputs_for_model
 
             calibration_seq_length = 256
             inputs = (
-                InputRecorder(
+                LMEvalInputRecorder(
                     tokenizer,
                     calibration_seq_length,
                     prepare_inputs_for_model,
@@ -736,7 +736,7 @@ def main(
                     ["wikitext"],
                     1,
                 )
-                .get_inputs()[0]
+                .get_recorded_inputs()[0]
                 .values[0]
             )
             inputs = prepare_inputs_for_model(inputs)
